requirements.txt
# Core dependencies
alpaca-py>=0.39.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
websockets>=11.0.0
ta>=0.10.0  # Technical analysis library
python-telegram-bot>=20.0  # For Telegram notifications
schedule>=1.2.0  # For scheduling tasks

# Database
aiosqlite>=0.19.0
SQLAlchemy>=2.0.0
alembic>=1.12.0

# Caching and messaging
redis>=4.6.0
msgpack>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-timeout>=2.1.0

# Code quality
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
flake8>=6.1.0
bandit>=1.7.5

# Data processing and visualization
plotly>=5.17.0
python-dateutil>=2.8.2
pytz>=2023.3

# System monitoring
statsd>=4.0.0
datadog>=0.45.0

# Type stubs
types-aiohttp>=3.9.0
types-PyYAML>=6.0.0
types-redis>=4.6.0
types-requests>=2.31.0
types-setuptools>=68.2.0
types-urllib3>=1.26.25

# Security
pyjwt[crypto]>=2.8.0
passlib>=1.7.4
bcrypt>=4.0.1
python-jose>=3.3.0

# Documentation
Sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Plugin system
PyYAML>=6.0.0
importlib-metadata>=6.0.0
importlib-resources>=5.0.0
